// types.ts
export enum SimulationType {
  PENDULUM = 'Simple Pendulum',
  PROJECTILE = 'Projectile Motion',
  SPRING_MASS = 'Spring-Mass System',
  ELECTRIC_FIELD = 'Electric Field',
  DOUBLE_SLIT = 'Double-Slit Experiment',
  SPECIAL_RELATIVITY = 'Special Relativity',
  QUANTUM_TUNNELING = 'Quantum Tunneling',
  RLC_CIRCUIT = 'RLC Circuit',
}